lab3/WPA2KeyExtractionService/api/requirements.txt
flask
flask-restful
pika
python-logging-loki
prometheus-client
